requirements.txt
aiohttp==3.8.3
aioprometheus==23.3.0
alembic==1.9.3
asyncpg==0.27.0
bcrypt==4.0.1
click==8.1.3
cryptography==3.4.7
elastic-apm==6.18.0
environs==9.3.1
fastapi==0.89.1
fastapi_filter==0.5.4
greenlet==2.0.2
ipython~=8.8.0
orjson==3.8.3
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
python-json-logger==2.0.4
requests==2.28.2
SQLAlchemy==1.4.46
sqlalchemy-utils==0.38.3
starlette~=0.22.0
typer~=0.7.0
uvicorn==0.15.0
uvloop==0.17.0
yarl~=1.8.2

-r requirements-dev.txt
